+++ changelog.md
@@ -17,6 +17,7 @@ Saves from 8.0.0 are compatible with 8.1.0
 ## Features/Improvements
 
 * **[Engine]** Support for DCS 2.8.6.41363, including F-15E support.
+* **[UI]** Flight loadout/properties tab is now scrollable.
 
 ## Fixes